+++ lib/CM5Receiver/CM5RightConnector.tsx
@@ -1,8 +1,119 @@
 import type { ChipProps } from "@tscircuit/props"
 import { CM5Connector } from "./CM5Connector"
 
+const pinLabelsRightSide = {
+  pin1: ["GND_1"],
+  pin2: ["GND_2"],
+  pin3: ["Ethernet_Pair3_P"],
+  pin4: ["Ethernet_Pair1_P"],
+  pin5: ["Ethernet_Pair3_N"],
+  pin6: ["Ethernet_Pair1_N"],
+  pin7: ["GND_3"],
+  pin8: ["GND_4"],
+  pin9: ["Ethernet_Pair2_N"],
+  pin10: ["Ethernet_Pair0_N"],
+  pin11: ["Ethernet_Pair2_P"],
+  pin12: ["Ethernet_Pair0_P"],
+  pin13: ["GND_5"],
+  pin14: ["GND_6"],
+  pin15: ["Ethernet_nLED3"],
+  pin16: ["Fan_Tacho"],
+  pin17: ["Ethernet_nLED2"],
+  pin18: ["Ethernet_SYNC_OUT"],
+  pin19: ["Fan_PWM"],
+  pin20: ["EEPROM_nWP"],
+  pin21: ["LED_nACT"],
+  pin22: ["GND_7"],
+  pin23: ["GND_8"],
+  pin24: ["GPIO26"],
+  pin25: ["GPIO21"],
+  pin26: ["GPIO19"],
+  pin27: ["GPIO20"],
+  pin28: ["GPIO13"],
+  pin29: ["GPIO16"],
+  pin30: ["GPIO6"],
+  pin31: ["GPIO12"],
+  pin32: ["GND_9"],
+  pin33: ["GND_10"],
+  pin34: ["GPIO5"],
+  pin35: ["ID_SC"],
+  pin36: ["ID_SD"],
+  pin37: ["GPIO7"],
+  pin38: ["GPIO11"],
+  pin39: ["GPIO8"],
+  pin40: ["GPIO9"],
+  pin41: ["GPIO25"],
+  pin42: ["GND_11"],
+  pin43: ["GND_12"],
+  pin44: ["GPIO10"],
+  pin45: ["GPIO24"],
+  pin46: ["GPIO22"],
+  pin47: ["GPIO23"],
+  pin48: ["GPIO27"],
+  pin49: ["GPIO18"],
+  pin50: ["GPIO17"],
+  pin51: ["GPIO15"],
+  pin52: ["GND_13"],
+  pin53: ["GND_14"],
+  pin54: ["GPIO4"],
+  pin55: ["GPIO14"],
+  pin56: ["GPIO3"],
+  pin57: ["SD_CLK"],
+  pin58: ["GPIO2"],
+  pin59: ["GND_15"],
+  pin60: ["GND_16"],
+  pin61: ["SD_DAT3"],
+  pin62: ["SD_CMD"],
+  pin63: ["SD_DAT0"],
+  pin64: ["SD_DAT5"],
+  pin65: ["GND_17"],
+  pin66: ["GND_18"],
+  pin67: ["SD_DAT1"],
+  pin68: ["SD_DAT4"],
+  pin69: ["SD_DAT2"],
+  pin70: ["SD_DAT7"],
+  pin71: ["GND_19"],
+  pin72: ["SD_DAT6"],
+  pin73: ["SD_VDD_OVERRIDE"],
+  pin74: ["GND_20"],
+  pin75: ["SD_PWR_ON"],
+  pin76: ["VBAT"],
+  pin77: ["V5_IN_1"],
+  pin78: ["GPIO_VREF"],
+  pin79: ["V5_IN_2"],
+  pin80: ["SCL0"],
+  pin81: ["V5_IN_3"],
+  pin82: ["SDA0"],
+  pin83: ["V5_IN_4"],
+  pin84: ["CM5_V3_3_OUT_1"],
+  pin85: ["V5_IN_5"],
+  pin86: ["CM5_V3_3_OUT_2"],
+  pin87: ["V5_IN_6"],
+  pin88: ["CM5_V1_8_OUT_1"],
+  pin89: ["WL_nDisable"],
+  pin90: ["CM5_V1_8_OUT_2"],
+  pin91: ["BT_nDisable"],
+  pin92: ["PWR_Button"],
+  pin93: ["nRPIBOOT"],
+  pin94: ["CC1"],
+  pin95: ["LED_nPWR"],
+  pin96: ["CC2"],
+  pin97: ["CAM_GPIO0"],
+  pin98: ["GND_21"],
+  pin99: ["PMIC_Enable"],
+  pin100: ["CAM_GPIO1"],
+}
+
+export type CM5RightPinLabels =
+  (typeof pinLabelsRightSide)[keyof typeof pinLabelsRightSide][number]
+
 export const CM5RightConnector = (props: {
   connections?: ChipProps["connections"]
 }) => (
-  <CM5Connector side="right" name="CM5_RIGHT" connections={props.connections} />
+  <CM5Connector
+    side="right"
+    name="CM5_RIGHT"
+    pinLabels={pinLabelsRightSide}
+    connections={props.connections}
+  />
 )
